added followers and following
