+++ pub-critic/src/app/components/UserInfo/UserInfo.module.scss
@@ -1,0 +1,29 @@
+.container {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  h1 {
+    margin-top: 2vw;
+    font-size: 4vw;
+    font-weight: 900;
+  }
+
+  .info {
+    flex-direction: column;
+    display: flex;
+    align-items: space-between;
+    width: 100%;
+    margin-top: 2vw;
+    font-size: 2vw;
+  }
+
+  .label {
+    font-size: 2vw;
+    font-weight: 500;
+    margin-top: 2vw;
+  }
+}
